complete toolbar
